new toolchain: sdcc+binutils for mcs51

+++ res/data/builtin_headers/lint_sdcc.h
@@ -49,5 +49,5 @@ typedef volatile unsigned short __sfr16;
 typedef volatile unsigned long __sfr32;
 
 #if defined(__SDCC_GNU_AS)
-    extern void __delay_cycles(unsigned long x);
+extern void __delay_cycles(unsigned int x);
 #endif
